google oauth

+++ my-app/src/components/XscrollBox.jsx
@@ -1,12 +1,125 @@
+// import React, { useEffect } from 'react';
+// import styles from './xscrollBox.module.css';
+// import { Link } from 'react-router-dom';
+// import { TbMovie } from "react-icons/tb";
+// import { FaRegBookmark } from "react-icons/fa";
+// import { IoMdBookmark } from "react-icons/io";
+// // import { Tooltip, Button } from '@chakra-ui/react';
+// import { useDispatch, useSelector } from 'react-redux';
+// import { addMovieBookmark, getMoviesBookmark, removeBookmark, removeMovieBookmark, requestAction, requestTrendingDataFetch } from '../redux/action';
+
+// function XscrollBox(props) {
+//     const moviesBookmark = useSelector((store) => store.moviesBookmark);
+//     const dispatch = useDispatch();
+//     const loading = useSelector((store) => store.isLoading);
+//     const error = useSelector((store) => store.isError);
+//     const data = useSelector((store) => store.trendingData);
+
+//     useEffect(() => {
+//         dispatch(requestAction());
+//         dispatch(requestTrendingDataFetch(props.api));
+//         dispatch(getMoviesBookmark())
+//     }, [dispatch, props.api]);
+
+//     const addBookmarkHandler = (ID) => {
+//         dispatch(addMovieBookmark({ "movieId": ID }));
+//     };
+
+//     const removeBookmarkHandler = (ID) => {
+//         dispatch(removeMovieBookmark({ "movieId": ID }));
+//     };
+
+
+//     return (
+//         <div className={styles.box1}>
+//             <div className={styles.title}>
+//                 Trending
+//             </div>
+//             <div className={styles.trending}>
+//                 {
+//                     loading ? <div className={styles.spinner}></div> : error ? <h1>Something went wrong</h1> :
+//                         data.length > 0 && data.map((el) => (
+//                             <div className={styles.box} key={el.id}>
+//                                 <div className={styles.cover}>
+//                                     <div className={styles.bookmark}>
+//                                         {/* <Tooltip label='' placement='top' bg='teal.600'> */}
+//                                         <button style={{ border: "none" }}>
+//                                             {moviesBookmark.includes(el.id) ?
+//                                                 <IoMdBookmark onClick={() => removeBookmarkHandler(el.id)} className={styles.booked} /> :
+//                                                 <FaRegBookmark onClick={() => addBookmarkHandler(el.id)} />
+//                                             }
+//                                         </button>
+//                                         {/* </Tooltip> */}
+//                                     </div>
+//                                     <Link to={`movieDetail/${el.id}`}>
+//                                         <img src={`https://image.tmdb.org/t/p/w300${el.backdrop_path}`} alt="" className={styles.img} />
+//                                         <p className={styles['josefin-sans']}>
+//                                             {el.release_date.slice(0, 4)} <span className={styles.dot}></span>
+//                                             {/* <TbMovie className={styles.small} /> */}
+//                                             <span className={styles.media}>Movie</span>
+//                                             <span className={styles.dot}></span>
+//                                             <span className={styles.media}>{el.adult ? "E" : "PG"}</span>
+//                                         </p>
+//                                         <h2 className={styles['josefin-sans']}>{el.title.length > 22 ? el.title.slice(0, 22) + "..." : el.title}</h2>
+//                                     </Link>
+//                                 </div>
+//                             </div>
+//                         ))
+//                 }
+//                 {
+//                     loading ? "" : error ? "" :
+//                         data.length > 0 && data.map((el) => (
+//                             <div className={styles.box} key={el.id}>
+//                                 <div className={styles.cover}>
+//                                     <div className={styles.bookmark}>
+//                                         {/* <Tooltip label='' placement='top' bg='teal.600'> */}
+//                                         <button style={{ border: "none" }}>
+//                                             {moviesBookmark.includes(el.id) ?
+//                                                 <IoMdBookmark onClick={() => removeBookmarkHandler(el.id)} className={styles.booked} /> :
+//                                                 <FaRegBookmark onClick={() => addBookmarkHandler(el.id)} />
+//                                             }
+//                                         </button>
+//                                         {/* </Tooltip> */}
+//                                     </div>
+//                                     <Link to={`movieDetail/${el.id}`}>
+//                                         <img src={`https://image.tmdb.org/t/p/w300${el.backdrop_path}`} alt="" className={styles.img} />
+//                                         <p className={styles['josefin-sans']}>
+//                                             {el.release_date.slice(0, 4)} <span className={styles.dot}></span>
+//                                             <TbMovie className={styles.small} />
+//                                             <span className={styles.media}>Movie</span>
+//                                             <span className={styles.dot}></span>
+//                                             <span className={styles.media}>{el.adult ? "E" : "PG"}</span>
+//                                         </p>
+//                                         <h2 className={styles['josefin-sans']}>
+//                                             {el.title.length > 22 ? el.title.slice(0, 22) + "..." : el.title}
+//                                         </h2>
+//                                     </Link>
+//                                 </div>
+//                             </div>
+//                         ))
+//                 }
+//             </div>
+//         </div>
+//     );
+// }
+
+// export default XscrollBox;
+
+
+
+
+
+
+
+
 import React, { useEffect } from 'react';
 import styles from './xscrollBox.module.css';
 import { Link } from 'react-router-dom';
 import { TbMovie } from "react-icons/tb";
 import { FaRegBookmark } from "react-icons/fa";
 import { IoMdBookmark } from "react-icons/io";
-// import { Tooltip, Button } from '@chakra-ui/react';
 import { useDispatch, useSelector } from 'react-redux';
-import {  addMovieBookmark, getMoviesBookmark, removeBookmark, removeMovieBookmark, requestAction, requestTrendingDataFetch } from '../redux/action';
+import { addMovieBookmark, getMoviesBookmark, removeMovieBookmark, requestAction, requestTrendingDataFetch } from '../redux/action';
 
 function XscrollBox(props) {
     const moviesBookmark = useSelector((store) => store.moviesBookmark);
@@ -14,21 +127,29 @@ function XscrollBox(props) {
     const loading = useSelector((store) => store.isLoading);
     const error = useSelector((store) => store.isError);
     const data = useSelector((store) => store.trendingData);
+    const isAuth = useSelector((store) => store.isAuth);
 
     useEffect(() => {
         dispatch(requestAction());
         dispatch(requestTrendingDataFetch(props.api));
-        dispatch(getMoviesBookmark())
-    }, [dispatch, props.api]);
+        if (isAuth) {
+            dispatch(getMoviesBookmark());
+        }
+    }, [dispatch, props.api, isAuth]);
 
     const addBookmarkHandler = (ID) => {
-        dispatch(addMovieBookmark({ "movieId": ID }));
+        if (isAuth) {
+            dispatch(addMovieBookmark({ "movieId": ID }));
+        } else {
+            alert('Please login to bookmark movies.');
+        }
     };
 
     const removeBookmarkHandler = (ID) => {
-        dispatch(removeMovieBookmark({ "movieId": ID }));
+        if (isAuth) {
+            dispatch(removeMovieBookmark({ "movieId": ID }));
+        }
     };
-
 
     return (
         <div className={styles.box1}>
@@ -37,66 +158,64 @@ function XscrollBox(props) {
             </div>
             <div className={styles.trending}>
                 {
-                loading ? <div className={styles.spinner}></div> : error ? <h1>Something went wrong</h1> :
-                    data.length > 0 && data.map((el) => (
-                        <div className={styles.box} key={el.id}>
-                            <div className={styles.cover}>
-                                <div className={styles.bookmark}>
-                                    {/* <Tooltip label='' placement='top' bg='teal.600'> */}
+                    loading ? <div className={styles.spinner}></div> : error ? <h1>Something went wrong</h1> :
+                        data.length > 0 && data.map((el) => (
+                            <div className={styles.box} key={el.id}>
+                                <div className={styles.cover}>
+                                    <div className={styles.bookmark}>
+                                        {/* <Tooltip label='' placement='top' bg='teal.600'> */}
                                         <button style={{ border: "none" }}>
-                                            {moviesBookmark.includes(el.id) ? 
-                                                <IoMdBookmark onClick={() => removeBookmarkHandler(el.id)} className={styles.booked} /> : 
+                                            {isAuth && moviesBookmark.includes(el.id) ?
+                                                <IoMdBookmark onClick={() => removeBookmarkHandler(el.id)} className={styles.booked} /> :
                                                 <FaRegBookmark onClick={() => addBookmarkHandler(el.id)} />
                                             }
                                         </button>
-                                    {/* </Tooltip> */}
+                                        {/* </Tooltip> */}
+                                    </div>
+                                    <Link to={`movieDetail/${el.id}`}>
+                                        <img src={`https://image.tmdb.org/t/p/w300${el.backdrop_path}`} alt="" className={styles.img} />
+                                        <p className={styles['josefin-sans']}>
+                                            {el.release_date.slice(0, 4)} <span className={styles.dot}></span>
+                                            {/* <TbMovie className={styles.small} /> */}
+                                            <span className={styles.media}>Movie</span>
+                                            <span className={styles.dot}></span>
+                                            <span className={styles.media}>{el.adult ? "E" : "PG"}</span>
+                                        </p>
+                                        <h2 className={styles['josefin-sans']}>{el.title.length > 22 ? el.title.slice(0, 22) + "..." : el.title}</h2>
+                                    </Link>
                                 </div>
-                                <Link to={`movieDetail/${el.id}`}>
-                                    <img src={`https://image.tmdb.org/t/p/w300${el.backdrop_path}`} alt="" className={styles.img} />
-                                    <p className={styles['josefin-sans']}>
-                                        {el.release_date.slice(0, 4)} <span className={styles.dot}></span> 
-                                        {/* <TbMovie className={styles.small} /> */}
-                                        <span className={styles.media}>Movie</span>
-                                        <span className={styles.dot}></span>
-                                        <span className={styles.media}>{el.adult ? "E" : "PG"}</span>
-                                    </p>
-                                    <h2 className={styles['josefin-sans']}>{el.title.length > 22 ? el.title.slice(0,22)+"..." : el.title }</h2>
-                                </Link>
                             </div>
-                        </div>
-                    ))
+                        ))
                 }
                 {
-                loading ? "" : error ? "" :
-                    data.length > 0 && data.map((el) => (
-                        <div className={styles.box} key={el.id}>
-                            <div className={styles.cover}>
-                                <div className={styles.bookmark}>
-                                    {/* <Tooltip label='' placement='top' bg='teal.600'> */}
+                    loading ? <div className={styles.spinner}></div> : error ? <h1>Something went wrong</h1> :
+                        data.length > 0 && data.map((el) => (
+                            <div className={styles.box} key={el.id}>
+                                <div className={styles.cover}>
+                                    <div className={styles.bookmark}>
+                                        {/* <Tooltip label='' placement='top' bg='teal.600'> */}
                                         <button style={{ border: "none" }}>
-                                            {moviesBookmark.includes(el.id) ? 
-                                                <IoMdBookmark onClick={() => removeBookmarkHandler(el.id)} className={styles.booked} /> : 
+                                            {isAuth && moviesBookmark.includes(el.id) ?
+                                                <IoMdBookmark onClick={() => removeBookmarkHandler(el.id)} className={styles.booked} /> :
                                                 <FaRegBookmark onClick={() => addBookmarkHandler(el.id)} />
                                             }
                                         </button>
-                                    {/* </Tooltip> */}
+                                        {/* </Tooltip> */}
+                                    </div>
+                                    <Link to={`movieDetail/${el.id}`}>
+                                        <img src={`https://image.tmdb.org/t/p/w300${el.backdrop_path}`} alt="" className={styles.img} />
+                                        <p className={styles['josefin-sans']}>
+                                            {el.release_date.slice(0, 4)} <span className={styles.dot}></span>
+                                            {/* <TbMovie className={styles.small} /> */}
+                                            <span className={styles.media}>Movie</span>
+                                            <span className={styles.dot}></span>
+                                            <span className={styles.media}>{el.adult ? "E" : "PG"}</span>
+                                        </p>
+                                        <h2 className={styles['josefin-sans']}>{el.title.length > 22 ? el.title.slice(0, 22) + "..." : el.title}</h2>
+                                    </Link>
                                 </div>
-                                <Link to={`movieDetail/${el.id}`}>
-                                    <img src={`https://image.tmdb.org/t/p/w300${el.backdrop_path}`} alt="" className={styles.img} />
-                                    <p className={styles['josefin-sans']}>
-                                        {el.release_date.slice(0, 4)} <span className={styles.dot}></span> 
-                                        <TbMovie className={styles.small} />
-                                        <span className={styles.media}>Movie</span>
-                                        <span className={styles.dot}></span>
-                                        <span className={styles.media}>{el.adult ? "E" : "PG"}</span>
-                                    </p>
-                                    <h2 className={styles['josefin-sans']}>
-                                       {el.title.length > 22 ? el.title.slice(0,22)+"..." : el.title }
-                                        </h2>
-                                </Link>
                             </div>
-                        </div>
-                    ))
+                        ))
                 }
             </div>
         </div>
@@ -104,8 +223,3 @@ function XscrollBox(props) {
 }
 
 export default XscrollBox;
-
-
-
-
-
